% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-finder.R
\name{approx_grad}
\alias{approx_grad}
\title{approximate gradient of functions}
\usage{
approx_grad(func, x, dx = .Machine$double.eps^(1/3))
}
\arguments{
\item{func}{function to take gradient of}

\item{x}{where the function to take gradient of}

\item{dx}{delta x}
}
\value{
gradient of function at input x
}
\description{
approximate gradient of functions using finite difference method
}
\details{
This function approximate gradient of functions using finite difference method, when we do not know the exact formula for the gradient
}
\examples{
approx_grad(function(x) log_likelihood(x, design, outcome), beta)

}
