% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{are_close}
\alias{are_close}
\title{Test two numeric values are close}
\usage{
are_close(v, w, abs_tol = 1e-06, rel_tol = 1e-06)
}
\arguments{
\item{v}{first numeric value}

\item{w}{second numeric value}

\item{abs_tol}{absolute tolerance}

\item{rel_tol}{relative tolerance}
}
\value{
TRUE if the difference between the two numeric values does not exceed absolute and relative tolerance, FALSE otherwise
}
\description{
Test two numeric values are close
}
\details{
This function check if two numeric value are close in terms of absolute and relative difference
}
\examples{
are_close(1, 1.01)

}
