% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiper_logit.R
\name{hiper_glm}
\alias{hiper_glm}
\title{fit GLM}
\usage{
hiper_glm(design, outcome, model = "linear", option = NULL)
}
\arguments{
\item{design}{design matrix}

\item{outcome}{outcome vector}

\item{par}{vector of parameter to estimate}

\item{noise_var}{optional error parameter}
}
\value{
hiper glm class object containing coefficients of the model
}
\description{
fit GLM of input design and outcome matrix
}
\details{
This function is fit a GLM model
}
\examples{
hiper_glm(design, outcome, model = 'linear', option = list(mle_solver = 'BFGS')

}
