% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiper_logit.R
\name{log_likelihood}
\alias{log_likelihood}
\title{log likelihood function (simplified) for linear model}
\usage{
log_likelihood(par, design, outcome, noise_var = 1)
}
\arguments{
\item{par}{vector of parameter to estimate}

\item{design}{design matrix}

\item{outcome}{outcome vector}

\item{noise_var}{optional error parameter}
}
\description{
log likelihood function (simplified) for linear model
}
