% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-finder.R
\name{log_likelihood_gradient}
\alias{log_likelihood_gradient}
\title{gradient of log likelihood function for linear model, taking derivative over beta}
\usage{
log_likelihood_gradient(par, design, outcome, noise_var = 1)
}
\arguments{
\item{par}{vector of parameter to estimate}

\item{design}{design matrix}

\item{outcome}{outcome vector}

\item{noise_var}{optional error parameter}
}
\value{
gradient of log likelihood function of linear model with with respect to betas
}
\description{
gradient of log likelihood function for linear model, taking derivative over beta
}
\details{
This function calculate the gradient of log likelihood function of linear model with respect to betas (betas/coefficients)
}
\examples{
log_likelihood_gradient(par, design, outcome)

}
