% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-func.R
\name{log_likelihood_linear}
\alias{log_likelihood_linear}
\title{log likelihood function for linear model}
\usage{
log_likelihood_linear(par, design, outcome, noise_var = 1)
}
\arguments{
\item{par}{vector of parameter to estimate}

\item{design}{design matrix}

\item{outcome}{outcome vector}

\item{noise_var}{optional error parameter}
}
\value{
log likelihood function for linear model
}
\description{
log likelihood function for linear model
}
\details{
This function calculate the log likelihood function of linear model simplified to only contain parts related to parameter (betas/coefficients)
}
\examples{
log_likelihood_linear(par, design, outcome)

}
