% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-func.R
\name{log_likelihood_logit}
\alias{log_likelihood_logit}
\title{log likelihood function for logistic model}
\usage{
log_likelihood_logit(par, design, outcome)
}
\arguments{
\item{par}{vector of parameter to estimate}

\item{design}{design matrix}

\item{outcome}{outcome vector}
}
\value{
log likelihood function for logistic model
}
\description{
log likelihood function for logistic model
}
\details{
This function calculate the log likelihood function of logistic model simplified to only contain parts related to parameter (betas/coefficients)
}
\examples{
log_likelihood_logit(par, design, outcome)

}
