% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-func.R
\name{log_likelihood_logit_hessian}
\alias{log_likelihood_logit_hessian}
\title{Hessian of log likelihood function for logistic model, taking double derivatives over beta}
\usage{
log_likelihood_logit_hessian(par, design, outcome)
}
\arguments{
\item{par}{vector of parameter to estimate}

\item{design}{design matrix}

\item{outcome}{outcome vector}
}
\value{
Hessian of log likelihood function of logistic model with with respect to betas
}
\description{
Hessian of log likelihood function for logistic model, taking double derivatives over beta
}
\details{
This function calculate the Hessian of log likelihood function of logistic model with respect to betas
}
\examples{
log_likelihood_logit_hessian(par, design, outcome)

}
