% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiper_logit.R
\name{mle_finder_inv}
\alias{mle_finder_inv}
\title{linear model mle finder using pseudoinverse method}
\usage{
mle_finder_inv(design, outcome)
}
\arguments{
\item{design}{design matrix}

\item{outcome}{outcome vector}
}
\description{
linear model mle finder using pseudoinverse method
}
