% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-finder.R
\name{mle_finder_logit_nr}
\alias{mle_finder_logit_nr}
\title{finding mle of logit by Newton Raphson}
\usage{
mle_finder_logit_nr(design, outcome, epsilon = 1e-08)
}
\arguments{
\item{design}{design matrix}

\item{outcome}{outcome vector}

\item{epsilon}{convergence threshold for log likelihood between two consecutive iterations}
}
\value{
mle coefficient of logit
}
\description{
finding mle of logit by Newton Raphson
}
\details{
This function calculate mle of logistic regression of Bernoulli response variable using Newton Raphson
}
\examples{
mle_finder_logit_nr(design, outcome)

}
