% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-finder.R
\name{mle_finder_optim}
\alias{mle_finder_optim}
\title{linear model mle}
\usage{
mle_finder_optim(design, outcome, func, grad, option)
}
\arguments{
\item{design}{design matrix}

\item{outcome}{outcome vector}
}
\value{
mle of linear model, approximated
}
\description{
linear model mle finder using optimization method
}
\details{
This function calculate mle of linear model using optimization method with a pre-specified solver
}
\examples{
mle_finder_optim(design, outcome, func, grad, option = list(mle_solver = 'BFGS'))
}
