% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-finder.R
\name{log_likelihood}
\alias{log_likelihood}
\title{log likelihood function}
\usage{
log_likelihood(par, design, outcome, noise_var = 1)
}
\arguments{
\item{par}{vector of parameter to estimate}

\item{design}{design matrix}

\item{outcome}{outcome vector}

\item{noise_var}{optional error parameter}
}
\value{
log likelihood function
}
\description{
log likelihood function for linear model, simplified
}
\details{
This function calculate the log likelihood function of linear model simplified to only contain parts related to betas
}
\examples{
log_likelihood(par, design, outcome)

}
