% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-finder.R
\name{mle_finder_inv}
\alias{mle_finder_inv}
\title{linear model mle finder, by pseudoinverse method}
\usage{
mle_finder_inv(design, outcome)
}
\arguments{
\item{design}{design matrix}

\item{outcome}{outcome vector}
}
\value{
mle of linear model
}
\description{
linear model mle finder, by pseudoinverse method
}
\details{
This function calculate mle of linear model using pseudoinverse
}
\examples{
mle_finder_inv(design, outcome)

}
