% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-finder.R
\name{mle_finder_optim}
\alias{mle_finder_optim}
\title{finding mle by optimization}
\usage{
mle_finder_optim(design, outcome, func, grad, option)
}
\arguments{
\item{design}{design matrix}

\item{outcome}{outcome vector}
}
\value{
mle of input likelihood (and score), approximated
}
\description{
finding mle by optimization
}
\details{
This function calculate mle of given likelihood (and score) function using optimization method with a pre-specified solver
}
\examples{
mle_finder_optim(design, outcome, func, grad, option = list(mle_solver = 'BFGS'))

}
