% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiper_logit.R
\name{mle_finder_optim}
\alias{mle_finder_optim}
\title{linear model mle finder using optimization method}
\usage{
mle_finder_optim(design, outcome, func, grad, option)
}
\arguments{
\item{design}{design matrix}

\item{outcome}{outcome vector}
}
\description{
linear model mle finder using optimization method
}
